Implementing SRP into code. backend methods for weather data will be in weather_utils.
